+++ os_info/src/info.rs
@@ -15,7 +15,7 @@ use super::{Bitness, Type, Version};
 /// use os_info;
 ///
 /// let info = os_info::get();
-/// println!("OS information: {}", info);
+/// println!("OS information: {info}");
 /// ```
 #[derive(Debug, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
@@ -26,7 +26,7 @@ pub struct Info {
     pub(crate) version: Version,
     /// Operating system edition.
     pub(crate) edition: Option<String>,
-    /// Operating system edition.
+    /// Operating system codename.
     pub(crate) codename: Option<String>,
     /// Operating system architecture in terms of how many bits compose the basic values it can deal
     /// with. See `Bitness` for details.
@@ -35,6 +35,8 @@ pub struct Info {
     // system model info
     pub(crate) model: Option<String>,
 
+    /// Processor architecture.
+    pub(crate) architecture: Option<String>,
 }
 
 impl Info {
@@ -51,6 +53,7 @@ impl Info {
     /// assert_eq!(None, info.edition());
     /// assert_eq!(None, info.codename());
     /// assert_eq!(Bitness::Unknown, info.bitness());
+    /// assert_eq!(None, info.architecture());
     /// ```
     pub fn unknown() -> Self {
         Self {
@@ -59,8 +62,8 @@ impl Info {
             edition: None,
             codename: None,
             bitness: Bitness::Unknown,
-            model: None
-
+            model: None,
+            architecture: None,
         }
     }
 
@@ -78,6 +81,7 @@ impl Info {
     /// assert_eq!(None, info.edition());
     /// assert_eq!(None, info.codename());
     /// assert_eq!(Bitness::Unknown, info.bitness());
+    /// assert_eq!(None, info.architecture());
     /// ```
     pub fn with_type(os_type: Type) -> Self {
         Self {
@@ -156,6 +160,18 @@ impl Info {
     /// get system model info
     pub fn model(&self) -> Option<&str> {
         self.model.as_ref().map(String::as_ref)
+
+    /// Returns operating system architecture.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use os_info::Info;
+    ///
+    /// let info = Info::unknown();
+    /// assert_eq!(None, info.architecture());
+    pub fn architecture(&self) -> Option<&str> {
+        self.architecture.as_ref().map(String::as_ref)
     }
 }
 
@@ -172,10 +188,10 @@ impl Display for Info {
             write!(f, " {}", self.version)?;
         }
         if let Some(ref edition) = self.edition {
-            write!(f, " ({})", edition)?;
+            write!(f, " ({edition})")?;
         }
         if let Some(ref codename) = self.codename {
-            write!(f, " ({})", codename)?;
+            write!(f, " ({codename})")?;
         }
         write!(f, " [{}]", self.bitness)
     }
@@ -194,26 +210,32 @@ mod tests {
         assert_eq!(None, info.edition());
         assert_eq!(None, info.codename());
         assert_eq!(Bitness::Unknown, info.bitness());
+        assert_eq!(None, info.architecture());
     }
 
     #[test]
     fn with_type() {
         let types = [
             Type::Redox,
+            Type::Alpaquita,
             Type::Alpine,
             Type::Amazon,
             Type::Android,
             Type::Arch,
+            Type::Artix,
             Type::CentOS,
             Type::Debian,
             Type::Emscripten,
             Type::EndeavourOS,
             Type::Fedora,
+            Type::Gentoo,
             Type::Linux,
             Type::Macos,
             Type::Manjaro,
             Type::Mariner,
             Type::NixOS,
+            Type::OpenCloudOS,
+            Type::openEuler,
             Type::openSUSE,
             Type::OracleLinux,
             Type::Pop,
@@ -271,6 +293,14 @@ mod tests {
             ),
             (
                 Info {
+                    os_type: Type::Artix,
+                    version: Version::Rolling(None),
+                    ..Default::default()
+                },
+                "Artix Linux Rolling Release [unknown bitness]",
+            ),
+            (
+                Info {
                     os_type: Type::Manjaro,
                     version: Version::Rolling(Some("2020.05.24".to_owned())),
                     ..Default::default()
@@ -308,7 +338,8 @@ mod tests {
                     edition: Some("edition".to_owned()),
                     codename: Some("codename".to_owned()),
                     bitness: Bitness::X64,
-                    model: None
+                    model: None,
+                    architecture: Some("architecture".to_owned()),
                 },
                 "Mac OS 10.2.0 (edition) (codename) [64-bit]",
             ),