+++ os_info/src/os_type.rs
@@ -6,6 +6,8 @@ use std::fmt::{self, Display, Formatter};
 #[allow(non_camel_case_types, clippy::upper_case_acronyms)]
 #[non_exhaustive]
 pub enum Type {
+    /// Alpaquita Linux (<https://bell-sw.com/alpaquita-linux/>).
+    Alpaquita,
     /// Alpine Linux (<https://en.wikipedia.org/wiki/Alpine_Linux>).
     Alpine,
     /// Amazon Linux AMI (<https://en.wikipedia.org/wiki/Amazon_Machine_Image#Amazon_Linux_AMI>).
@@ -14,6 +16,8 @@ pub enum Type {
     Android,
     /// Arch Linux (<https://en.wikipedia.org/wiki/Arch_Linux>).
     Arch,
+    /// Artix Linux (<https://en.wikipedia.org/wiki/Artix_Linux>).
+    Artix,
     /// CentOS (<https://en.wikipedia.org/wiki/CentOS>).
     CentOS,
     /// Debian (<https://en.wikipedia.org/wiki/Debian>).
@@ -28,19 +32,27 @@ pub enum Type {
     Fedora,
     /// FreeBSD (<https://en.wikipedia.org/wiki/FreeBSD>).
     FreeBSD,
-    /// Linux based operating system (<https://en.wikipedia.org/wiki/Linux>).
+    /// Garuda Linux (<https://en.wikipedia.org/wiki/Garuda_Linux>)
+    Garuda,
+    /// Gentoo Linux (<https://en.wikipedia.org/wiki/Gentoo_Linux>).
+    Gentoo,
+    /// HardenedBSD (https://hardenedbsd.org/).
     HardenedBSD,
-    /// HardenedBSD (https://hardenedbsd.org/)
+    /// Illumos (https://en.wikipedia.org/wiki/Illumos).
+    Illumos,
+    /// Linux based operating system (<https://en.wikipedia.org/wiki/Linux>).
     Linux,
+    /// Mabox (<https://maboxlinux.org/>).
+    Mabox,
     /// Mac OS X/OS X/macOS (<https://en.wikipedia.org/wiki/MacOS>).
     Macos,
     /// Manjaro (<https://en.wikipedia.org/wiki/Manjaro>).
     Manjaro,
     /// Mariner (<https://en.wikipedia.org/wiki/CBL-Mariner>).
     Mariner,
-    /// Mint (<https://en.wikipedia.org/wiki/Linux_Mint>).
+    /// MidnightBSD (<https://en.wikipedia.org/wiki/MidnightBSD>).
     MidnightBSD,
-    /// MidnightBSD(<https://en.wikipedia.org/wiki/MidnightBSD>).
+    /// Mint (<https://en.wikipedia.org/wiki/Linux_Mint>).
     Mint,
     /// NetBSD (<https://en.wikipedia.org/wiki/NetBSD>).
     NetBSD,
@@ -48,6 +60,10 @@ pub enum Type {
     NixOS,
     /// OpenBSD (<https://en.wikipedia.org/wiki/OpenBSD>).
     OpenBSD,
+    /// OpenCloudOS (<https://www.opencloudos.org>).
+    OpenCloudOS,
+    /// openEuler (<https://en.wikipedia.org/wiki/EulerOS>).
+    openEuler,
     /// openSUSE (<https://en.wikipedia.org/wiki/OpenSUSE>).
     openSUSE,
     /// Oracle Linux (<https://en.wikipedia.org/wiki/Oracle_Linux>).
@@ -83,10 +99,15 @@ impl Default for Type {
 impl Display for Type {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         match *self {
+            Type::Alpaquita => write!(f, "Alpaquita Linux"),
             Type::Alpine => write!(f, "Alpine Linux"),
             Type::Amazon => write!(f, "Amazon Linux AMI"),
             Type::Arch => write!(f, "Arch Linux"),
+            Type::Artix => write!(f, "Artix Linux"),
             Type::DragonFly => write!(f, "DragonFly BSD"),
+            Type::Garuda => write!(f, "Garuda Linux"),
+            Type::Gentoo => write!(f, "Gentoo Linux"),
+            Type::Illumos => write!(f, "illumos"),
             Type::Macos => write!(f, "Mac OS"),
             Type::MidnightBSD => write!(f, "Midnight BSD"),
             Type::Mint => write!(f, "Linux Mint"),
@@ -95,7 +116,7 @@ impl Display for Type {
             Type::Redhat => write!(f, "Red Hat Linux"),
             Type::RedHatEnterprise => write!(f, "Red Hat Enterprise Linux"),
             Type::SUSE => write!(f, "SUSE Linux Enterprise Server"),
-            _ => write!(f, "{:?}", self),
+            _ => write!(f, "{self:?}"),
         }
     }
 }
@@ -112,16 +133,20 @@ mod tests {
     #[test]
     fn display() {
         let data = [
+            (Type::Alpaquita, "Alpaquita Linux"),
             (Type::Alpine, "Alpine Linux"),
             (Type::Amazon, "Amazon Linux AMI"),
             (Type::Android, "Android"),
             (Type::Arch, "Arch Linux"),
+            (Type::Artix, "Artix Linux"),
             (Type::CentOS, "CentOS"),
             (Type::Debian, "Debian"),
             (Type::DragonFly, "DragonFly BSD"),
             (Type::Emscripten, "Emscripten"),
             (Type::EndeavourOS, "EndeavourOS"),
             (Type::Fedora, "Fedora"),
+            (Type::Garuda, "Garuda Linux"),
+            (Type::Gentoo, "Gentoo Linux"),
             (Type::FreeBSD, "FreeBSD"),
             (Type::Linux, "Linux"),
             (Type::Macos, "Mac OS"),